Refactor utils + fix sum methods
